+++ cloudpunch/slave/ping.py
@@ -4,8 +4,12 @@ import logging
 import time
 
 import cloudpunch.utils.config as cpc
+import cloudpunch.utils.metrics as metrics
 
 from threading import Thread
+
+METRIC_NAME = 'cloudpunch.ping'
+LATENCY_METRIC = '%s.latency' % METRIC_NAME
 
 
 class CloudPunchTest(Thread):
@@ -13,6 +17,8 @@ class CloudPunchTest(Thread):
     def __init__(self, config):
         self.config = config
         self.final_results = []
+        if self.config['metrics']['enable']:
+            self.metric = metrics.Metrics(self.config['metrics'])
         super(CloudPunchTest, self).__init__()
 
     def run(self):
@@ -34,13 +40,10 @@ class CloudPunchTest(Thread):
         # Configuration setup
         if self.config['server_client_mode']:
             server_ip = self.config['match_ip']
-        elif 'target' in self.config['ping']:
-            server_ip = self.config['ping']['target']
         else:
-            raise ConfigError('Missing ping target server')
+            server_ip = self.config['ping']['target']
         duration = str(self.config['ping']['duration'])
 
-        results = []
         logging.info('Starting ping command to server %s for %s seconds', server_ip, duration)
         ping = subprocess.Popen(['ping', '-c', duration, server_ip],
                                 stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
@@ -48,40 +51,19 @@ class CloudPunchTest(Thread):
         for line in iter(ping.stdout.readline, ''):
             latency = re.findall(r'time=(\d+\.\d+)', line)
             now = int(time.time())
-            if latency:
-                latency = float(latency[0])
-                # Over time results
-                if self.config['overtime_results']:
-                    self.final_results.append({
-                        'time': now,
-                        'latency': latency
-                    })
-                # Summary results
-                else:
-                    results.append(latency)
-            # Ping failed
-            elif 'Request timeout' in line and self.config['overtime_results']:
-                self.final_results.append({
-                    'time': now,
-                    'latency': 0
-                })
+            latency = float(latency[0]) if latency else 0
+            self.final_results.append({
+                'time': now,
+                'latency': latency
+            })
+            if self.config['metrics']['enable']:
+                self.metric.send_metric(LATENCY_METRIC, latency, now)
 
         ping.stdout.close()
 
         # Send back summary if not over time
         if not self.config['overtime_results']:
-            try:
-                self.final_results = {
-                    'latency': sum(results) / len(results)
-                }
-            except ZeroDivisionError:
-                self.final_results = {
-                    'latency': -1
-                }
-
-
-class ConfigError(Exception):
-
-    def __init__(self, message):
-        super(ConfigError, self).__init__(message)
-        self.message = message
+            latency = [d['latency'] for d in self.final_results]
+            self.final_results = {
+                'latency': sum(latency) / len(latency)
+            }