+++ CHANGELOG.md
@@ -3,8 +3,9 @@ A list of notable changes will be documented here
 
 ## 2.0.0 - 2017-xx-xx
 ### Added
+- All official tests now include sending metrics to Kafka. Currently, the only supported format is influxdb. See documentation for more information
+
 - Take over existing resources
-- Add open source metrics
 - Add testing for swift
 - Added a new `-i` option for the run command. This allows connection to an existing master to rerun tests
 
@@ -18,6 +19,9 @@ A list of notable changes will be documented here
 - The jmeter test now has time as the current time rather than the test duration to match other tests
 - The run command now uses `-f FORMAT` instead of `--yaml`. Yaml is now the default option
 - The run command uses `-b FLAVOR_FILE` instead of `-f` for providing a flavor breakdown
+
+### Removed
+- Admin mode has been removed, CloudPunch will now only use the current user and project
 - The run command no longer has `--split`. Supplying a second OpenRC file will enable split mode
 - The run command no longer has `--reuse`. Use the new `-i` option as a replacement